remove uneeded execution-context reference
